🚧  Fix ruff preview level but easy to fix errors (#102)
* 🚧  Fix ruff preview level but easy to fix errors
  (FURB110, E302, RUF046, E303, COM812, DOC402)

* 🚨  Fix tiny mypy error & update github actions beacuase of numba 0.61 is released.

* 🤖 Add dtype

+++ src/arpes/deep_learning/interpret.py
@@ -167,13 +167,13 @@ class Interpretation:
             if isinstance(n_items, tuple):
                 layout = n_items
             else:
-                n_rows = int(math.ceil(n_items**0.5))
+                n_rows = math.ceil(n_items**0.5)
                 layout = (n_rows, n_rows)
 
             items = self.top_losses()[:n_items]
         else:
             n_items = len(items)
-            n_rows = int(math.ceil(n_items**0.5))
+            n_rows = math.ceil(n_items**0.5)
             layout = (n_rows, n_rows)
 
         assert isinstance(n_items, int)
